neural_future_backend/npc/serializers.py
from rest_framework import serializers

from npc.models import NPC, Question


class NPCSerializer(serializers.ModelSerializer):
    class Meta:
        model = NPC
        fields = '__all__'


class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = ('id', 'question')
